get admin with pagination, remove and edit functionality

+++ apps/frontend/src/app/dashboard/create-admin/create-admin.component.ts
@@ -7,7 +7,7 @@ import {
   StateModel,
   User
 } from '@covid19-helpline/models';
-import { FormBuilder, FormGroup, Validators } from '@angular/forms';
+import { FormBuilder, FormControl, FormGroup, Validators } from '@angular/forms';
 import { AuthService } from '../../shared/services/auth.service';
 import { Subject } from 'rxjs';
 import { debounceTime } from 'rxjs/operators';
@@ -16,6 +16,7 @@ import { CityService } from '../../shared/services/state-city/city.service';
 import { GeneralService } from '../../shared/services/general.service';
 import { UserService } from '../../shared/services/user/user.service';
 import { Router } from '@angular/router';
+import { NzModalService } from 'ng-zorro-antd';
 
 @Component({
   selector: 'covid19-helpline-create-admin',
@@ -39,6 +40,8 @@ export class CreateAdminComponent implements OnInit {
   public allData: User[] = [];
   public isGettingDataInProcess:boolean;
 
+  public formTitle: string = 'Create Admin';
+
 
   public filterRequest: GetAllAdminUsersRequestModel = {
     count : 15,
@@ -47,6 +50,7 @@ export class CreateAdminComponent implements OnInit {
     sort:'',
     sortBy: 'desc'
   }
+  public isVisibleConfirmRemove: boolean;
 
   public modelChangedState = new Subject<string>();
   public modelChangedCity = new Subject<string>();
@@ -57,7 +61,8 @@ export class CreateAdminComponent implements OnInit {
               private _authService: AuthService,
               private _generalService: GeneralService,
               private _userService: UserService,
-              private _router: Router) { }
+              private _router: Router,
+              private modalService: NzModalService) { }
 
   ngOnInit() {
 
@@ -115,22 +120,42 @@ export class CreateAdminComponent implements OnInit {
   }
 
 
-  public initForm() {
+  public initForm(isCancel?: boolean) {
+
+    this.formTitle = 'Create Admin';
+
     this.adminForm = this.FB.group(
       {
+        id: [null],
         firstName: [null, [Validators.required]],
         lastName: [null, [Validators.required]],
-        city: [null],
-        state: [null],
-        cityId: [null],
-        stateId: [null],
-        emailId: [null],
-        mobileNumber: [null, Validators.required],
+        city: [null, [Validators.required]],
+        cityId: [null, [Validators.required]],
+        state: [null, [Validators.required]],
+        stateId: [null, [Validators.required]],
+        mobileNumber: [null, [Validators.required]],
       }
     );
 
-    this.selectedCity = null;
-    this.selectedState = null;
+    if(!isCancel) {
+
+    }
+
+    this.isRequestInProcess = false;
+    this.selectedCity = this._generalService.user.city;
+    this.selectedState = this._generalService.user.state;
+
+    if(this._generalService.user.state) {
+      this.adminForm.get('stateId').patchValue(this._generalService.user.state.id);
+      this.adminForm.get('state').patchValue(this._generalService.user.state.name);
+    }
+
+    if(this._generalService.user.city) {
+      this.adminForm.get('cityId').patchValue(this._generalService.user.city.id);
+      this.adminForm.get('city').patchValue(this._generalService.user.city.name);
+    }
+
+
 
   }
 
@@ -176,14 +201,35 @@ export class CreateAdminComponent implements OnInit {
 
   }
 
+
+  enableEdit(user:User) {
+    this.formTitle = 'Edit Admin';
+    this.adminForm.patchValue(user);
+    this.adminForm.get('cityId').patchValue(user.city.id);
+    this.adminForm.get('stateId').patchValue(user.state.id);
+  }
+
+
+
   async removeAdmin(id: string) {
 
     try {
 
-      const json = { id : id };
-      this.isRequestInProcess = true;
-      await this._userService.deleteAdmin(json).toPromise();
-      this.isRequestInProcess = false;
+      return this.modalService.confirm({
+        nzTitle: 'Are you sure delete?',
+        nzContent: '',
+        nzOnOk: () =>
+          new Promise(async (resolve, reject) => {
+            const json = { id: id };
+            this.isRequestInProcess = true;
+            const data = await this._userService.deleteAdmin(json).toPromise();
+            this.isRequestInProcess = false;
+              if(data && data.data) {
+                setTimeout(Math.random() > 0.5 ? resolve : reject, 10);
+              }
+            return true;
+          }).catch(() => console.log('Oops errors!'))
+      });
 
     } catch (e) {
       this.isRequestInProcess = false;
@@ -206,6 +252,11 @@ export class CreateAdminComponent implements OnInit {
       this.isGettingDataInProcess = false;
     }
 
+  }
+
+  public pageChanged(index: number) {
+    this.filterRequest.page = index;
+    this.getAllAdmin();
   }
 
 }
